Rename Logger.h as LogDestination.h
Rename of class will be a separate commit
