+++ srcStatic/Log/LogDestination.h
