+++ srcStatic/op2extStatic.vcxproj
@@ -167,7 +167,7 @@
     <ClInclude Include="IniFile.h" />
     <ClInclude Include="LocalResource.h" />
     <ClInclude Include="Log.h" />
-    <ClInclude Include="Log/Logger.h" />
+    <ClInclude Include="Log\LogDestination.h" />
     <ClInclude Include="Log/LoggerDebug.h" />
     <ClInclude Include="Log/LoggerDistributor.h" />
     <ClInclude Include="Log/LoggerFile.h" />
